+++ src/index.js
@@ -3,3 +3,4 @@ export { default } from './extendRoutine';
 export { default as createExtendedRoutine } from './createExtendedRoutine';
 export { default as createCustomRoutine } from './createCustomRoutine';
 export { default as createSocketRoutine } from './createSocketRoutine';
+export { default as createRoutines } from './createRoutines';