+++ test/createRoutines.js
@@ -1,0 +1,3 @@
+import { createRoutines } from '../src';
+
+createRoutines({});